Unify some files
