+++ packages/react-jss/src/utils/sheets.js
@@ -1,10 +1,11 @@
 // @flow
 import warning from 'tiny-warning'
 import {getDynamicStyles, type StyleSheetFactoryOptions} from 'jss'
-import type {Context, Styles} from '../types'
+import type {Context, DynamicRules, Styles} from '../types'
 import {getManager} from './managers'
 import {jss as defaultJss} from '../jss'
-import {addMetaForSheet} from './sheets-meta'
+import {addMetaForSheet, getMetaForSheet} from './sheets-meta'
+import type {StyleSheet} from 'jss'
 
 interface Options<Theme> {
   context: Context;
@@ -45,7 +46,7 @@ function getSheetOptions<Theme>(options: Options<Theme>, link: boolean) {
   }
 }
 
-function createSheet<Theme>(options: Options<Theme>) {
+function createStaticSheet<Theme>(options: Options<Theme>) {
   const manager = getManager(options.context, options.index)
   const existingSheet = manager.get(options.theme)
 
@@ -69,4 +70,55 @@ function createSheet<Theme>(options: Options<Theme>) {
   return sheet
 }
 
-export {createSheet}
+const removeDynamicRules = (sheet: ?StyleSheet, rules: ?DynamicRules) => {
+  if (!sheet || !rules) {
+    return
+  }
+
+  // Loop over each dynamic rule and remove the dynamic rule
+  // We can't just remove the whole sheet as this has all of the rules for every component instance
+  for (const key in rules) {
+    sheet.deleteRule(rules[key].key)
+  }
+}
+
+const updateDynamicRules = (data: any, sheet: ?StyleSheet, rules: ?DynamicRules) => {
+  if (!sheet || !rules) {
+    return
+  }
+
+  // Loop over each dynamic rule and update it
+  // We can't just update the whole sheet as this has all of the rules for every component instance
+  for (const key in rules) {
+    // $FlowFixMe
+    sheet.update(rules[key].key, data)
+  }
+}
+
+const addDynamicRules = (sheet: ?StyleSheet): ?DynamicRules => {
+  if (!sheet) {
+    return undefined
+  }
+
+  const meta = getMetaForSheet(sheet)
+
+  if (!meta) {
+    return undefined
+  }
+
+  const rules: DynamicRules = {}
+
+  // Loop over each dynamic rule and add it to the stylesheet
+  for (const key in meta.dynamicStyles) {
+    const name = `${key}-${meta.dynamicRuleCounter++}`
+    const rule = sheet.addRule(name, meta.dynamicStyles[key])
+
+    if (rule) {
+      rules[key] = rule
+    }
+  }
+
+  return rules
+}
+
+export {addDynamicRules, updateDynamicRules, removeDynamicRules, createStaticSheet}
